LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigate-event/navigate-navigation-back-same-document-in-iframe-expected.txt

CONSOLE MESSAGE: Error: assert_equals: expected (object) null but got (undefined) undefined


Harness Error (FAIL), message = Error: assert_equals: expected (object) null but got (undefined) undefined

PASS navigate event for navigation.back() - same-document in an iframe
